fix: 404 darkmode
