+++ ui/src/App.tsx
@@ -1,6 +1,8 @@
 import loadable from '@loadable/component';
 import nightwind from 'nightwind/helper';
+import { useEffect, useState } from 'react';
 import { Helmet } from 'react-helmet';
+import { useSelector } from 'react-redux';
 import { createHashRouter, RouterProvider } from 'react-router-dom';
 import { SWRConfig } from 'swr';
 import ErrorLayout from './app/ErrorLayout';
@@ -10,10 +12,12 @@ import Flag from './app/flags/Flag';
 import NewFlag from './app/flags/NewFlag';
 import Layout from './app/Layout';
 import NotFoundLayout from './app/NotFoundLayout';
+import { selectTheme } from './app/preferences/preferencesSlice';
 import NewSegment from './app/segments/NewSegment';
 import Segment from './app/segments/Segment';
 import SessionProvider from './components/SessionProvider';
 import { request } from './data/api';
+import { Theme } from './types/Preferences';
 const Flags = loadable(() => import('./app/flags/Flags'));
 const Segments = loadable(() => import('./app/segments/Segments'));
 const Console = loadable(() => import('./app/console/Console'));
@@ -127,6 +131,25 @@ const fetcher = async (uri: String) => {
 };
 
 export default function App() {
+  const theme = useSelector(selectTheme);
+  const [systemPrefersDark, setSystemPrefersDark] = useState(
+    window.matchMedia('(prefers-color-scheme: dark)').matches
+  );
+
+  useEffect(() => {
+    window
+      .matchMedia('(prefers-color-scheme: dark)')
+      .addEventListener('change', (e) => setSystemPrefersDark(e.matches));
+  }, []);
+
+  useEffect(() => {
+    if (theme === Theme.SYSTEM) {
+      nightwind.enable(systemPrefersDark);
+    } else {
+      nightwind.enable(theme === Theme.DARK);
+    }
+  }, [theme, systemPrefersDark]);
+
   return (
     <>
       <Helmet>
